requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database drivers
SQLAlchemy==2.0.23
pyodbc==5.0.1

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1

# Development
pytest==7.4.3
pytest-flask==1.3.0

